aws/credentials/ec2rolecreds: Moved EC2 Role Provier out of Credentials
In order to be use the EC2 Metadata service without import cycles the
EC2RoleProvider needs to be outside of the base credentials package.

Migration:
If your application was importing the `credentials.EC2RoleProvider` you'll
just need to update to use ec2rolecreds.EC2RoleProvider instead. If using
the `credentials.NewEC2RoleCredentials` you can now call `ec2rolecreds.New()`
