+++ aws/config.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws/credentials"
+	"github.com/aws/aws-sdk-go/aws/credentials/ec2rolecreds"
 )
 
 // DefaultChainCredentials is a Credentials which will find the first available
@@ -17,7 +18,7 @@ var DefaultChainCredentials = credentials.NewChainCredentials(
 	[]credentials.Provider{
 		&credentials.EnvProvider{},
 		&credentials.SharedCredentialsProvider{Filename: "", Profile: ""},
-		&credentials.EC2RoleProvider{ExpiryWindow: 5 * time.Minute},
+		&ec2rolecreds.EC2RoleProvider{ExpiryWindow: 5 * time.Minute},
 	})
 
 // The default number of retries for a service. The value of -1 indicates that
